Fix Redis client language seeping into Ruby API-land
Kredis isn't about being a thin Ruby wrapper on top of Redis, but about exposing higher level data structures in a way that's more natural to Ruby and it's APIs.

So if we're adding a Hash type, the API to mirror is Ruby's native Hash API, not what the underlying Redis command calls may be.

Thus del should be delete, same as the many other types wrapping del as reset, clear, remove etc.